Extract AopProps class
