+++ src/CodeGen.php
@@ -4,36 +4,25 @@ declare(strict_types=1);
 
 namespace Ray\Aop;
 
-use Doctrine\Common\Annotations\AnnotationException;
-use Doctrine\Common\Annotations\Reader as DoctrineReader;
 use PhpParser\BuilderFactory;
 use PhpParser\Node\Identifier;
 use PhpParser\Node\Name;
 use PhpParser\Node\Stmt;
 use PhpParser\Node\Stmt\Class_;
-use PhpParser\Node\Stmt\Property;
 use PhpParser\Parser;
-use Ray\ServiceLocator\ServiceLocator;
 use ReflectionClass;
 
 use function array_merge;
 use function assert;
 use function implode;
-use function serialize;
 
 final class CodeGen implements CodeGenInterface
 {
-    /** @var Parser */
-    private $parser;
-
     /** @var BuilderFactory */
     private $factory;
 
     /** @var CodeGenMethod */
     private $codeGenMethod;
-
-    /** @var DoctrineReader */
-    private $reader;
 
     /** @var AopClassName */
     private $aopClassName;
@@ -41,20 +30,19 @@ final class CodeGen implements CodeGenInterface
     /** @var VisitorFactory */
     private $visitoryFactory;
 
-    /**
-     * @throws AnnotationException
-     */
+    /** @var AopProps */
+    private $aopProps;
+
     public function __construct(
         Parser $parser,
         BuilderFactory $factory,
         AopClassName $aopClassName
     ) {
-        $this->parser = $parser;
         $this->factory = $factory;
         $this->codeGenMethod = new CodeGenMethod($parser);
-        $this->reader = ServiceLocator::getReader();
         $this->aopClassName = $aopClassName;
-        $this->visitoryFactory = new VisitorFactory($this);
+        $this->visitoryFactory = new VisitorFactory($parser);
+        $this->aopProps = new AopProps($factory);
     }
 
     /**
@@ -67,7 +55,7 @@ final class CodeGen implements CodeGenInterface
         $visitor = ($this->visitoryFactory)($sourceClass);
         assert($visitor->class instanceof Class_);
         $methods = $this->codeGenMethod->getMethods($bind, $visitor);
-        $propStms = $this->getAopProps($sourceClass);
+        $propStms = ($this->aopProps)($sourceClass);
         $classStm = $visitor->class;
         $newClassName = ($this->aopClassName)((string) $visitor->class->name, $bind->toString(''));
         $classStm->name = new Identifier($newClassName);
@@ -88,73 +76,6 @@ final class CodeGen implements CodeGenInterface
     }
 
     /**
-     * @param ReflectionClass<object> $class
-     */
-    private function getClassAnnotation(ReflectionClass $class): string
-    {
-        $classAnnotations = $this->reader->getClassAnnotations($class);
-
-        return serialize($classAnnotations);
-    }
-
-    /**
-     * @param ReflectionClass<object> $class
-     *
-     * @return Property[]
-     */
-    private function getAopProps(ReflectionClass $class): array
-    {
-        $pros = [];
-        $pros[] = $this->factory
-            ->property('bind')
-            ->makePublic()
-            ->getNode();
-
-        $pros[] =
-            $this->factory->property('bindings')
-                ->makePublic()
-                ->setDefault([])
-                ->getNode();
-
-        $pros[] = $this->factory
-            ->property('methodAnnotations')
-            ->setDefault($this->getMethodAnnotations($class))
-            ->makePublic()
-            ->getNode();
-        $pros[] = $this->factory
-            ->property('classAnnotations')
-            ->setDefault($this->getClassAnnotation($class))
-            ->makePublic()
-            ->getNode();
-        $pros[] = $this->factory
-            ->property('isAspect')
-            ->makePrivate()
-            ->setDefault(true)
-            ->getNode();
-
-        return $pros;
-    }
-
-    /**
-     * @param ReflectionClass<object> $class
-     */
-    private function getMethodAnnotations(ReflectionClass $class): string
-    {
-        $methodsAnnotation = [];
-        $methods = $class->getMethods();
-        foreach ($methods as $method) {
-            $annotations = $this->reader->getMethodAnnotations($method);
-            if ($annotations === []) {
-                continue;
-            }
-
-            $methodsAnnotation[$method->name] = $annotations;
-        }
-
-        return serialize($methodsAnnotation);
-    }
-
-    /**
      * @return string|null
      */
     private function getNamespace(CodeVisitor $source)
@@ -163,10 +84,5 @@ final class CodeGen implements CodeGenInterface
         $ns = implode('\\', $parts);
 
         return $ns ? $ns : null;
-    }
-
-    public function getParser(): Parser
-    {
-        return $this->parser;
     }
 }
